Organized Machine Learning Project/requirements.txt
torch>=1.9.0
pyYAML>=5.4.1
pytest=8.3.4